Move testdata under src/main dir Fix exception message check in VidModelSelectorTest Add IdVariable class
