+++ src/main/kotlin/org/wfanet/virtualpeople/core/selector/resourcekey/BUILD.bazel
@@ -6,7 +6,7 @@ package(default_visibility = [
 
 kt_jvm_library(
     name = "resource_key",
-    srcs = glob(["*Key.kt"]),
+    srcs = glob(["*Key.kt"]) + ["IdVariable.kt"],
     deps = [
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
     ],
