+++ src/main/kotlin/org/wfanet/virtualpeople/core/selector/testing/testdata/model_line_01.textproto
