+++ src/main/kotlin/org/wfanet/virtualpeople/core/selector/testing/testdata/model_line_02.textproto
