+++ src/main/kotlin/org/wfanet/virtualpeople/core/selector/testing/testdata/model_rollout_01.textproto
