+++ src/main/kotlin/org/wfanet/virtualpeople/core/selector/testing/testdata/model_rollout_02.textproto
