+++ src/main/kotlin/org/wfanet/virtualpeople/core/selector/testing/testdata/model_rollout_without_rollout_period_02.textproto
