+++ src/test/kotlin/org/wfanet/virtualpeople/core/selector/BUILD.bazel
@@ -4,11 +4,12 @@ kt_jvm_test(
     name = "VidModelSelectorTest",
     srcs = ["VidModelSelectorTest.kt"],
     data = [
-        "//src/test/proto/wfa/virtual_people/core/selector/test_data:selector_test_data",
+        "//src/main/kotlin/org/wfanet/virtualpeople/core/selector/testing/testdata:selector_test_data",
     ],
     test_class = "org.wfanet.virtualpeople.core.selector.VidModelSelectorTest",
     deps = [
         "//src/main/kotlin/org/wfanet/virtualpeople/core/selector",
+        "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/org/junit",
         "@wfa_common_jvm//imports/kotlin/kotlin/test",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/testing",
